pyproject.toml
[tool.poetry]
name = "apinonnopy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
dj-database-url = "^1.1.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
